+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
-### Added
+### Upcoming
 - Globalsentiment page Table
 - Sentiment Page
 - PSS Tags
+- Influence and Reach Page
+- All output numbers with 2 decimal places
 
+
+## [1.0.2.2] 2016-07-19
+## Added
+- Output Top Opinions (5 HardCored - Server Side only)
 
 ## [1.0.2.1] - 2016-07-19
 ### Fixed